+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.18.0] - 2024-07-05
+### Added
+- Enable deletion of a property in an INI document with a given mode ([#69](https://github.com/oliverzick/Delizious-Ini/issues/69))
+  - Fail mode: Throw section not found exception if the section does not exist, throw property not found exception if the section exists but the property to delete does not exist
+  - Ignore mode: Silently ignore if the section or the property to delete does not exist
+
 ## [0.17.0] - 2024-06-30
 ### Changed
 - Improve performance by superseding exception-based approach by strategy-based approach ([#70](https://github.com/oliverzick/Delizious-Ini/issues/70))